+++ packages/pragma/src/pragma/cli/commands/__init__.py
@@ -1,1 +1,0 @@
-"""One file per subcommand."""