+++ packages/pragma/src/pragma/report/__init__.py
@@ -1,1 +1,0 @@
-"""Post-Implementation Log: span aggregation + human / JSON reporting."""